Move hook into setup.py
Might as well use the file if it's just going to be sitting there right?
